+++ packages/react-microsurveys-editor/src/lib/components/dialog/StyleEditor/StyleEditor.tsx
@@ -4,7 +4,7 @@ import { Controller, useFormContext } from 'react-hook-form';
 import PaletteEditor from '../PaletteEditor/PaletteEditor';
 
 import { IconSun, IconMoon } from '@tabler/icons';
-
+import { createTheme, Theme } from '@mui/material/styles';
 import Grid from '@mui/material/Grid';
 import ToggleButtonGroup from '@mui/material/ToggleButtonGroup';
 import ToggleButton from '@mui/material/ToggleButton';
@@ -12,13 +12,21 @@ import ToggleButton from '@mui/material/ToggleButton';
 export interface StyleEditorProps {}
 
 export function StyleEditor(props: StyleEditorProps) {
-  const { control } = useFormContext<Form>();
+  const { control, watch } = useFormContext<Form>();
+  const mode = watch('settings.dialog.palette.mode');
+
+  const theme = createTheme({
+    palette: {
+      mode: mode,
+    },
+  });
 
   return (
     <Grid sx={{ flexGrow: 1 }} container spacing={2} direction="column">
       <Grid item>
         <Controller
           name={`settings.dialog.palette.mode`}
+          defaultValue={theme.palette.mode}
           control={control}
           render={({ field: { onChange, onBlur, value, ref } }) => (
             <ToggleButtonGroup
@@ -40,69 +48,105 @@ export function StyleEditor(props: StyleEditorProps) {
       <Grid item>
         <Controller
           name={`settings.dialog.palette.background.paper`}
-          defaultValue={'#fff'}
+          defaultValue={theme.palette.background.paper}
           control={control}
           rules={{ required: true }}
-          render={({ field }) => (
-            <PaletteEditor label="Background" {...field} />
+          render={({ field: { ref, onChange, value } }) => (
+            <PaletteEditor
+              label="Background"
+              inputRef={ref}
+              onChange={onChange}
+              value={value}
+              defaultValue={theme.palette.background.paper}
+            />
           )}
         />
       </Grid>
       <Grid item>
         <Controller
           name={`settings.dialog.palette.primary.main`}
+          defaultValue={theme.palette.primary.main}
           control={control}
           rules={{ required: true }}
-          render={({ field }) => (
-            <PaletteEditor label="Primary Button" {...field} />
+          render={({ field: { ref, onChange, value } }) => (
+            <PaletteEditor
+              label="Primary Button"
+              inputRef={ref}
+              onChange={onChange}
+              value={value}
+              defaultValue={theme.palette.primary.main}
+            />
           )}
         />
       </Grid>
       <Grid item>
         <Controller
           name={`settings.dialog.palette.secondary.main`}
+          defaultValue={theme.palette.secondary.main}
           control={control}
           rules={{ required: true }}
-          render={({ field }) => (
-            <PaletteEditor label="Cancel Button" {...field} />
+          render={({ field: { ref, onChange, value } }) => (
+            <PaletteEditor
+              label="Cancel Button"
+              inputRef={ref}
+              onChange={onChange}
+              value={value}
+              defaultValue={theme.palette.secondary.main}
+            />
           )}
         />
       </Grid>
       <Grid item>
         <Controller
           name={`settings.dialog.palette.text.primary`}
+          defaultValue={theme.palette.text.primary}
           control={control}
           rules={{ required: true }}
-          render={({ field }) => (
-            <PaletteEditor label="Primary Text" {...field} />
+          render={({ field: { ref, onChange, value } }) => (
+            <PaletteEditor
+              label="Primary Text"
+              inputRef={ref}
+              onChange={onChange}
+              value={value}
+              defaultValue={theme.palette.text.primary}
+            />
           )}
         />
       </Grid>
       <Grid item>
         <Controller
           name={`settings.dialog.palette.text.secondary`}
+          defaultValue={theme.palette.text.secondary}
           control={control}
           rules={{ required: true }}
-          render={({ field }) => (
-            <PaletteEditor label="Secondary Text" {...field} />
+          render={({ field: { ref, onChange, value } }) => (
+            <PaletteEditor
+              label="Secondary Text"
+              inputRef={ref}
+              onChange={onChange}
+              value={value}
+              defaultValue={theme.palette.text.secondary}
+            />
           )}
         />
       </Grid>
       <Grid item>
         <Controller
           name={`settings.dialog.palette.error.main`}
+          defaultValue={theme.palette.error.main}
           control={control}
           rules={{ required: true }}
-          render={({ field }) => <PaletteEditor label="Error" {...field} />}
+          render={({ field: { ref, onChange, value } }) => (
+            <PaletteEditor
+              label="Error"
+              inputRef={ref}
+              onChange={onChange}
+              value={value}
+              defaultValue={theme.palette.error.main}
+            />
+          )}
         />
       </Grid>
-
-      {/* <Controller
-        name={`settings.styles.palette.text.primary`}
-        control={control}
-        rules={{ required: true }}
-        render={({ field }) => <PaletteEditor label="text" {...field} />}
-      /> */}
     </Grid>
   );
 }